+++ src/main/java/codesquad/springcafe/controller/HomeController.java
@@ -1,7 +1,7 @@
 package codesquad.springcafe.controller;
 
 import codesquad.springcafe.model.Article;
-import codesquad.springcafe.repository.ArticleRepository;
+import codesquad.springcafe.repository.article.ArticleRepository;
 import java.util.List;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
